requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
pandas
openpyxl
faiss-cpu
python-dotenv
sentence-transformers
flask
flask-cors
reportlab
weasyprint
python-docx
streamlit
requests